feat: add combobox
